refactor: my last refactoring in this pull request
