+++ src/app/mainPage/MainPage.scss
