+++ src/app/sectionPage/SectionPage.scss
