+++ src/app/store.ts
@@ -1,6 +1,6 @@
 import { Action, configureStore, ThunkAction } from '@reduxjs/toolkit';
 import counterReducer from 'features/counter/counterSlice';
-import sectorsReducer from 'app/sectors/sectorsSlice';
+import sectorsReducer from 'features/sectors/sectorsSlice';
 
 export const store = configureStore({
   reducer: {
