+++ src/features/sectors/sectorsSlice.ts
@@ -1,41 +1,15 @@
 import { createSlice /* , PayloadAction */ } from '@reduxjs/toolkit';
 import type { RootState } from 'app/store';
-
+import * as t from 'types';
 import { PAGES_PER_SECTOR, SECTOR_COLORS } from '../../constants';
 
-export interface SectorPage {
-  sectorNum: number;
-  pageNum: number;
-}
-
-export interface SectorPageVisibility extends SectorPage {
-  visible: boolean;
-}
-
-export interface Page {
-  key: number;
-  title: string;
-  url: string;
-  show: boolean;
-}
-
-export interface Sector {
-  key: number;
-  title: string;
-  color: string;
-  pages: Array<Page>;
-}
-
-interface SectorsState {
-  sectors: Array<Sector>;
-}
-
-function generatePages(sectorNum: number, color: string): Array<Page> {
-  const pages: Array<Page> = [];
-  for (let i = 1; i <= PAGES_PER_SECTOR; i += 1) {
+/* Страницы нумеруются с нуля! */
+function generatePages(sectorNum: number, color: string): Array<t.Page> {
+  const pages: Array<t.Page> = [];
+  for (let i = 0; i < PAGES_PER_SECTOR; i += 1) {
     pages.push({
       key: i,
-      title: `Страница ${i}`,
+      title: `Страница ${i + 1}`,
       url: `/section/${sectorNum}/${i}/${encodeURIComponent(color)}`,
       show: true,
     });
@@ -43,18 +17,17 @@ function generatePages(sectorNum: number, color: string): Array<Page> {
   return pages;
 }
 
-const initialState: SectorsState = {
-  sectors: SECTOR_COLORS.map((color, index) => {
-    return {
-      key: index,
-      title: `Раздел ${index + 1}`,
-      color,
-      get pages() {
-        return generatePages(index, color);
-      },
-    };
-  }),
-};
+/* Разделы нумеруются с нуля! */
+const initialState: t.SectorsState = SECTOR_COLORS.map((color, index) => {
+  return {
+    key: index,
+    title: `Раздел ${index + 1}`,
+    color,
+    get pages() {
+      return generatePages(index, color);
+    },
+  };
+});
 
 export const sectorsSlice = createSlice({
   name: 'sectors',
@@ -62,11 +35,13 @@ export const sectorsSlice = createSlice({
   reducers: {
     /*
     setSectorPageVisibility: (state, action: PayloadAction<SectorPageVisibility>) => {
-      return state.sectors.map((item, index) => {
-        if (index !== action.payload.sectorNum) {
+      return state.map((item) => {
+        if (item.key !== action.payload.sectorNum) {
           // This isn't the item we care about - keep it as-is
-          return item;
+          return { ...item, pages: item.pages.map((el) => el) };
         }
+
+        return { ...item, pages: item.pages.map((el) => el) };
 
         // Otherwise, this is the one we want - return an updated value
         return {
@@ -77,22 +52,21 @@ export const sectorsSlice = createSlice({
               : { ...page, show: action.payload.visible }
           ),
         };
-      });
-    }, */
+      }); */
   },
 });
 
 // export const { setSectorPageVisibility } = sectorsSlice.actions;
 
-export const selectSectors = (state: RootState) => state.sectors.sectors;
+export const selectSectors = (state: RootState) => state.sectors;
 
 export const selectAdjacentPages = (
   state: RootState,
-  requestData: SectorPage
-): Array<Page | undefined> => {
+  requestData: t.SectorPage
+): Array<t.Page | undefined> => {
   let prevPage;
   let nextPage;
-  const sector = state.sectors.sectors.find((item) => item.key === requestData.sectorNum);
+  const sector = state.sectors.find((item) => item.key === requestData.sectorNum);
   if (sector) {
     for (let index = 0; index < sector.pages.length; index += 1) {
       const page = sector.pages[index];
