+++ src/features/sectorsPage/SectorsPage.tsx
@@ -11,7 +11,8 @@ import {
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 import { makeStyles } from '@material-ui/core/styles';
 import { useSelector } from 'react-redux';
-import { selectSectors, Sector } from 'app/sectors/sectorsSlice';
+import { selectSectors } from 'features/sectors/sectorsSlice';
+import * as t from 'types';
 
 import './SectorsPage.scss';
 
@@ -33,7 +34,7 @@ const useStyles = makeStyles((theme) => ({
 export default function SectionsPage(): JSX.Element {
   const classes = useStyles();
 
-  const sectors: Array<Sector> = useSelector(selectSectors);
+  const sectors: Array<t.Sector> = useSelector(selectSectors);
 
   return (
     <div className={classes.root}>
